OMS-N: Ensure unique type IDs when parsing overlays [12/12]
Always add a block to the idmap for each type in the overlay package
even if there are no matching resources for that type. This ensures that
ResTable::parsePackage will always get a unique ID for each type.

An alternative implementation would be to skip non-matching types when
parsing the overlay package, but that would prevent resolving references
to resources in XML.

Bug: 31052947

Change-Id: I059f2e8848dff685b5160fbd7ae20b8dfc64b1e5